requirements.txt
flask
python-telegram-bot
setting_command.py
setting_constant.py